requirements.txt
addict~=2.4.0
diffusers~=0.14.0
gradio~=3.25.0
huggingface_hub
matplotlib~=3.7.1
numpy~=1.24.2
onnxruntime~=1.14.1
opencv_python
Pillow~=9.2.0
pycocotools~=2.0.6
PyYAML~=6.0
requests~=2.25.1
setuptools~=67.6.0
supervision~=0.3.2
termcolor~=2.2.0
timm~=0.6.13
torch~=2.0.0+cu118
torchvision~=0.15.1+cu118
transformers~=4.26.0
yapf~=0.32.0

flask~=2.2.3
groundingdino~=0.1.0
opencv-python~=4.7.0.72